+++ src/Models/LwcAdminZone.php
@@ -14,4 +14,33 @@ class LwcAdminZone extends Model
      * @var array
      */
     protected $guarded = ['id'];
+
+    // RELATIONS
+
+    public function alternateNames()
+    {
+        return $this->hasMany('PaulinTrognon\LaravelWorldCities\Models\LwcAdminZoneAlternateName');
+    }
+
+    // Scopes
+
+    public function scopeAdm1($query)
+    {
+        $query->where('type', 'adm1');
+    }
+
+    public function scopeAdm2($query)
+    {
+        $query->where('type', 'adm2');
+    }
+
+    public function scopeAdm3($query)
+    {
+        $query->where('type', 'adm3');
+    }
+
+    public function scopeAdm4($query)
+    {
+        $query->where('type', 'adm4');
+    }
 }