+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## v2.2.0
+
+- Add select2 selection plugin support
+- Fix relationships
+
+### v2.1.1
+
+- Fix Model Name Error
+
+## v2.1.0
+
+- Add alternative names
+
 # v2.0.0
 
 - Add administrative subdivisions into `lwc_admin_zones` table